requirements.txt
flask
python-dotenv
pdfplumber
python-docx
pandas
google-genai
werkzeug
